+++ example/src/main/java/com/mercadopago/sdk/android/example/extensions/FloatExt.kt
@@ -1,10 +1,9 @@
 package com.mercadopago.sdk.android.example.extensions
 
-import java.math.BigDecimal
 import java.text.NumberFormat
 import java.util.Locale
 
-fun BigDecimal.toCurrencyString(locale: Locale = Locale.getDefault()): String {
+fun Float.toCurrencyString(locale: Locale = Locale.getDefault()): String {
     val format = NumberFormat.getCurrencyInstance(locale)
     return format.format(this)
 }
